+++ resources/js/app.js
@@ -595,6 +595,7 @@ Vue.component('file-upload', VueUploadComponent);
 // Extended Components
 import GlobalSearch from './components/GlobalSearch.vue';
 import EntitySearch from './components/EntitySearch.vue';
+import EntityTypeSearch from './components/EntityTypeSearch.vue';
 import LabelSearch from './components/LabelSearch.vue';
 import AttributeSearch from './components/AttributeSearch.vue';
 import CsvTable from './components/CsvTable.vue';
@@ -615,6 +616,7 @@ import ErrorModal from './components/Error.vue';
 
 Vue.component('global-search', GlobalSearch);
 Vue.component('entity-search', EntitySearch);
+Vue.component('entity-type-search', EntityTypeSearch);
 Vue.component('label-search', LabelSearch);
 Vue.component('attribute-search', AttributeSearch);
 Vue.component('csv-table', CsvTable);